Autonomous mode, pathfinding, stuff
- Added autonomous mode and three fully functional Steamworks
Autonomouses.
- Auto modes are written just like normal FRC commands.
- When there is only one controller, multiple "fake robots" can be added
and controlled individually.
- Started adding A* Pathfinding, for when AI robots are added.
- Added FakeButtons.
- Fixed a bug where gears could be picked up when they shouldn't be.

+++ core/src/ryan/game/bcnlib_pieces/Motor.java
@@ -2,13 +2,13 @@ package ryan.game.bcnlib_pieces;
 
 public class Motor {
 
-    double power;
+    float power;
 
-    public void setPower(double p) {
+    public void setPower(float p) {
         power = p;
     }
 
-    public double getPower() {
+    public float getPower() {
         return power;
     }
 }
